refactor: Added misc trpc router
