+++ apps/api/src/fsplit/routers/misc-router.ts
@@ -1,6 +1,22 @@
 import { router } from "../trpc";
-import { contactUs } from "../actions/misc";
+import {
+  completeVerification,
+  contactUs,
+  getPasswordResetToken,
+  getVerificationToken,
+  isVerified,
+  resendVerificationEmailByToken,
+  resetPassword,
+  sendResetPasswordEmail,
+} from "../actions/misc";
 
 export const miscRouter = router({
+  completeVerification,
   contactUs,
+  getPasswordResetToken,
+  getVerificationToken,
+  isVerified,
+  resendVerificationEmailByToken,
+  resetPassword,
+  sendResetPasswordEmail,
 });
