+++ apps/api/src/fsplit/routers/user-router.ts
@@ -15,35 +15,21 @@ import {
   createNewUser,
   changePassword,
   changePreferredCurrency,
-  resendVerificationEmailByToken,
-  completeVerification,
-  isVerified,
-  getVerificationToken,
-  sendResetPasswordEmail,
-  getPasswordResetToken,
-  resetPassword,
 } from "../actions/user";
 
 export const userRouter = router({
-  getAuthProfile,
-  getFriends,
-  getFriendRequests,
-  getGrossBalance,
-  getDebtWithFriend,
   acceptFriendRequest,
-  createNewUser,
   changePassword,
   changePreferredCurrency,
+  createNewUser,
   declineFriendRequest,
   editProfile,
   findFriends,
   findUsers,
+  getAuthProfile,
+  getDebtWithFriend,
+  getFriends,
+  getFriendRequests,
+  getGrossBalance,
   sendFriendRequest,
-  resendVerificationEmailByToken,
-  completeVerification,
-  isVerified,
-  getVerificationToken,
-  sendResetPasswordEmail,
-  getPasswordResetToken,
-  resetPassword,
 });
